fix(storage): tombstone re-gate on repaired refs; tier-move repair; chunk dedupe guard; entities sync scope
Review round 21:
- (Codex P1) writeMemory's resurrection gate is a re-runnable closure:
  when the post-persist repair changes the final entityRef, the gate
  re-runs against the NEW claimant before the rewrite (only ever ADDS
  blocking), so a tombstone keyed under the legacy claimant still blocks.
  The chunk/frontmatter tombstone paths never key on entityRef, so no
  re-gate applies there.
- (Codex P1) writeMemoryFileAtomic (tier moves) repairs instead of
  detects — the source is unlinked after the move, so a parked mapping
  across the write had nothing left to reconcile from.
- (Codex P2) chunk repair runs only when THIS chunk persisted
  (written === id) — a tombstone-dedupe result returns an existing id
  without writing, and the unconditional repair could create a stray
  file from scratch.
- (Bugbot) entities/*.md offline-sync writes request reconciliation too
  (relationship targets are in the migration's rewrite scope).
- Ratchet: citation-marker hash-source helpers extracted to
  storage/citation-hash-source.ts (pure move; storage.ts 7061 -> 7030
  with real headroom); both new modules registered in the
  lifecycle-matrix manifest.
