+++ scripts/lifecycle-matrix/coverage.json
@@ -14,6 +14,7 @@
     "packages/remnic-core/src/storage/entity-canonical-id-migration-runner.ts",
     "packages/remnic-core/src/storage/entity-canonical-id-migration-adapter.ts",
     "packages/remnic-core/src/storage/entity-canonical-id-references.ts",
+    "packages/remnic-core/src/storage/citation-hash-source.ts",
     "packages/remnic-core/src/storage/memory-frontmatter-metadata.ts",
     "packages/remnic-core/src/storage/memory-migration-serialization.ts",
     "packages/remnic-core/src/storage/**",
@@ -148,6 +149,7 @@
     "packages/remnic-core/src/storage/entity-canonical-id-migration-runner.ts": "extraction-lifecycle",
     "packages/remnic-core/src/storage/entity-canonical-id-migration-adapter.ts": "extraction-lifecycle",
     "packages/remnic-core/src/storage/entity-canonical-id-references.ts": "extraction-lifecycle",
+    "packages/remnic-core/src/storage/citation-hash-source.ts": "extraction-lifecycle",
     "packages/remnic-core/src/storage/memory-frontmatter-metadata.ts": "extraction-lifecycle",
     "packages/remnic-core/src/storage/memory-migration-serialization.ts": "extraction-lifecycle",
     "packages/remnic-core/src/storage/tombstone-blocked-capture-index.ts": "tombstone-blocked-capture",
